* tp/Texinfo/XS/main/manipulate_tree.c (copy_associated_info) (copy_tree_internal, remove_associated_copy_info, parse_node_manual) (modify_tree, normalized_menu_entry_internal_node, first_menu_node), tp/Texinfo/XS/main/tree.c (destroy_associated_info) (destroy_source_mark_list, destroy_element_and_children) (remove_element_from_list, add_element_if_not_in_list) (replace_element_in_list, next_bracketed_or_word_agg) (split_delimiters, split_def_args, parse_def), tp/Texinfo/XS/structuring_transfo/structuring.c (insert_menu_comment_content, new_complete_node_menu): replace int by size_t in situations where the variable could reach a value outside of the int range and also when the index is better reprensented by a size_t.

+++ tp/Texinfo/XS/structuring_transfo/structuring.c
@@ -1790,7 +1790,7 @@ new_node_menu_entry (const ELEMENT *node, int use_sections)
 }
 
 static void
-insert_menu_comment_content (ELEMENT_LIST *element_list, int position,
+insert_menu_comment_content (ELEMENT_LIST *element_list, size_t position,
                    ELEMENT *inserted_element, int no_leading_empty_line)
 {
   ELEMENT *menu_comment = new_element (ET_menu_comment);
@@ -1862,7 +1862,7 @@ new_complete_node_menu (const ELEMENT *node, DOCUMENT *document,
       const char *normalized = lookup_extra_string (node, AI_key_normalized);
       if (normalized && !strcmp (normalized, "Top"))
         {
-          int content_index = 0;
+          size_t content_index = 0;
           int in_appendix = 0;
           for (i = 0; i < node_childs->number; i++)
             {
